Revert "Extract reserve helpers into dedicated module"
