+++ src/components/AttendanceTab.tsx
@@ -6,8 +6,7 @@ import ClientDetailsModal from "./clients/ClientDetailsModal";
 import ClientForm from "./clients/ClientForm";
 import ColumnSettings from "./ColumnSettings";
 import { compareValues, toggleSort } from "./tableUtils";
-import { fmtDate, todayISO, uid } from "../state/utils";
-import { isReserveArea } from "../state/reserve";
+import { fmtDate, isReserveArea, todayISO, uid } from "../state/utils";
 import { commitDBUpdate } from "../state/appState";
 import { buildGroupsByArea, clientRequiresManualRemainingLessons } from "../state/lessons";
 import { transformClientFormValues } from "./clients/clientMutations";
