+++ src/components/__tests__/ClientsTab.test.tsx
@@ -30,13 +30,6 @@ jest.mock('../../state/utils', () => ({
 
 }));
 
-jest.mock('../../state/reserve', () => ({
-  __esModule: true,
-  isReserveArea: jest.fn(() => false),
-  ensureReserveAreaIncluded: jest.fn(v => v),
-  RESERVE_AREA_NAME: 'резерв',
-}));
-
 import ClientsTab from '../ClientsTab';
 import { commitDBUpdate } from '../../state/appState';
 import { uid, todayISO, fmtMoney, fmtDate, parseDateInput, calcAgeYears, calcExperience, calcExperienceMonths } from '../../state/utils';
