+++ src/components/__tests__/LeadsTab.test.tsx
@@ -25,13 +25,6 @@ jest.mock('../../state/utils', () => ({
   RESERVE_AREA_NAME: 'резерв',
 }));
 
-jest.mock('../../state/reserve', () => ({
-  __esModule: true,
-  isReserveArea: () => false,
-  ensureReserveAreaIncluded: (areas: string[]) => areas,
-  RESERVE_AREA_NAME: 'резерв',
-}));
-
 import LeadsTab from '../LeadsTab';
 import QuickAddModal from '../QuickAddModal';
 import { commitDBUpdate } from '../../state/appState';
