+++ src/state/appState.ts
@@ -6,8 +6,7 @@ import { useToasts } from "../components/Toasts";
 import { doc, onSnapshot, setDoc } from "firebase/firestore";
 import { db as firestore, ensureSignedIn } from "../firebase";
 import { makeSeedDB } from "./seed";
-import { todayISO, uid } from "./utils";
-import { ensureReserveAreaIncluded } from "./reserve";
+import { ensureReserveAreaIncluded, todayISO, uid } from "./utils";
 import { DEFAULT_SUBSCRIPTION_PLAN, getSubscriptionPlanMeta } from "./payments";
 import { ensureReserveAreaIncluded } from "./areas";
 import type {
