+++ src/state/seed.ts
@@ -1,5 +1,4 @@
-import { rnd, uid, todayISO } from "./utils";
-import { RESERVE_AREA_NAME } from "./reserve";
+import { RESERVE_AREA_NAME, rnd, uid, todayISO } from "./utils";
 import type {
   Area,
   Client,
